Before testing the router...
